UI Updates
Buttons and text on the menus and game scene have been enlarged to help with readability. The canvases have been adapted to scale with the screen in order to help with these scaling matters. The only screen untouched is the team/name selection screen as tampering with it seemed to break it.